+++ src/Janus.Authentication/Maintenance/IMaintenanceStore.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Janus.Core;
 
 namespace Janus.Authentication.Maintenance;