+++ src/Janus.Storage/Authentication/Maintenance/MaintenanceEntryRecord.cs
@@ -1,5 +1,4 @@
 using System;
-using Janus.Authentication.Maintenance;
 using Janus.Core;
 
 namespace Janus.Storage.Authentication.Maintenance;